Finish up refactoring geary into ecs and spigot

+++ ecs/src/main/java/com/mineinabyss/geary/ecs/systems/TickingSystem.kt
@@ -10,6 +10,6 @@ import com.mineinabyss.geary.ecs.engine.forEach
  *
  * @see [Engine.forEach]
  */
-public abstract class TickingSystem(public val interval: Int = 1) {
+public abstract class TickingSystem(public val interval: Long = 1) {
     public abstract fun tick()
 }
